requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
Django==4.2.6
django-ckeditor==6.7.1
django-js-asset==2.2.0
idna==3.7
Jinja2==3.1.3
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.0
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
preshed==3.0.9
psycopg==3.1.18
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
thinc==8.2.3
threadpoolctl==3.5.0
tqdm==4.66.4
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
